iox-#13 add config file parser and some refactorings

+++ iceoryx_posh/include/iceoryx_posh/capro/service_description.hpp
@@ -37,7 +37,7 @@ static constexpr char AnyEventString[]{"65535"};
 static constexpr int32_t MAX_NUMBER_OF_CHARS = 64;
 
 /// @brief Scope of a service description
-enum class Scope : unsigned int
+enum class Scope : uint16_t
 {
     WORLDWIDE,
     INTERNAL,
